requirements.txt
gradio
requests
smolagents>=1.1.0
litellm
duckduckgo-search
ddgs
markdownify
youtube-transcript-api
opencv-python
python-dotenv
pandas
pillow